# homework_5/swagger-open-api (1).yml
openapi: 3.0.0
info:
  title: StreamSync AI API
  version: 1.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/none-9f0-b5c/CC-5/1.0.0
  - url: https://api.streamsync-ai.com/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: Authenticates a user and returns a JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Unauthorized
  /jobs:
    get:
      summary: Retrieves a list of past video processing jobs for the authenticated user
      responses:
        '200':
          description: A list of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    job_id:
                      type: string
                      format: uuid
                    source_url:
                      type: string
                      format: uri
                    target_language:
                      type: string
                    status:
                      type: string
                    created_at:
                      type: string
                      format: date-time
  /jobs/create:
    post:
      summary: Creates a new asynchronous video processing job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_url:
                  type: string
                  format: uri
                target_language:
                  type: string
                  maxLength: 2
      responses:
        '202':
          description: Job successfully accepted and added to the processing queue
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
                  status:
                    type: string
  /jobs/{job_id}/status:
    get:
      summary: Retrieves the current status of a processing job
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current job status returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    enum: [PENDING, PROCESSING, COMPLETED, FAILED]
                  progress_percentage:
                    type: integer
  /jobs/{job_id}/subtitles:
    get:
      summary: Downloads the generated subtitle file
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Subtitle file returned
          content:
            text/plain:
              schema:
                type: string
  /jobs/{job_id}:
    delete:
      summary: Deletes a processing job and its associated subtitle resources
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Job successfully deleted
        '404':
          description: Job not found